Added distance to reduce degression factor

+++ src/main/java/net/avatarrealms/minecraft/bending/model/BendingPlayer.java
@@ -545,7 +545,7 @@ public class BendingPlayer {
 		return bendings.get(type).getLastTime();
 	}
 	
-	public int getSpamHistory(BendingType type) {
+	public double getSpamHistory(BendingType type) {
 		return bendings.get(type).getSpamHistory();
 	}
